Refactor header guards and update ssize_t to int64_t in multiple files for consistency and improved type handling

+++ src/bfc.c
@@ -28,7 +28,7 @@ int main(int argc, char** argv)
     bfc_token_stream_t* tok_stream = nullptr;
 
     [[gnu::cleanup(bfc_jump_table_destroy)]]
-    ssize_t* jump_table = nullptr;
+    int64_t* jump_table = nullptr;
 
     [[gnu::cleanup(bfc_ir_destroy)]]
     bfc_ir_block_t* root_block = nullptr;
